requirements.txt
# Core dependencies for src modules
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0

# Web app
Flask>=3.0.0
